requirements_dev_and_testing.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.3
idna==3.3
iniconfig==1.1.1
multidict==5.2.0
mysql-connector==2.2.9
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
toml==0.10.2
typing-extensions==4.0.1
yarl==1.7.2
